requirements.txt
asgiref==3.7.2
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
dj-database-url==2.0.0
dj-rest-auth==2.1.11
Django==4.2.3
django-allauth==0.48.0
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-embed-video==1.4.9
django-googledrive-storage==1.6.0
django-js-asset==2.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
google-api-core==2.10.2
google-api-python-client==2.95.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
